fix skywire cli survey
