+++ cmd/skywire-cli/commands/survey/root.go
@@ -2,12 +2,18 @@
 package clisurvey
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
+	"net/http"
 	"os"
 
+	"github.com/skycoin/dmsg/pkg/disc"
+	"github.com/skycoin/dmsg/pkg/dmsg"
+	"github.com/skycoin/skycoin/src/cipher"
 	"github.com/spf13/cobra"
 
+	"github.com/skycoin/skywire-utilities/pkg/cmdutil"
 	"github.com/skycoin/skywire-utilities/pkg/logging"
 	"github.com/skycoin/skywire-utilities/pkg/skyenv"
 	"github.com/skycoin/skywire/cmd/skywire-cli/internal"
@@ -32,7 +38,7 @@ var (
 func init() {
 	surveyCmd.Flags().SortFlags = false
 	surveyCmd.Flags().StringVarP(&confPath, "config", "c", "", "optionl config file to use (i.e.: "+visorconfig.ConfigName+")")
-	surveyCmd.Flags().StringVar(&dmsgDisc, "dmsg-disc", skyenv.DmsgDiscAddr, "value of dmsg discovery")
+	surveyCmd.Flags().StringVarP(&dmsgDisc, "dmsg-disc", "D", skyenv.DmsgDiscAddr, "value of dmsg discovery")
 	//	surveyCmd.Flags().StringVarP(&confArg, "confarg", "C", "", "supply config as argument")
 	//	surveyCmd.Flags().BoolVarP(&stdin, "stdin", "n", false, "read config from stdin")
 	if _, err := os.Stat(visorconfig.SkywirePath + "/" + visorconfig.ConfigJSON); err == nil {
@@ -64,19 +70,8 @@ var surveyCmd = &cobra.Command{
 		if usr {
 			confPath = visorconfig.HomePath() + "/" + visorconfig.ConfigName
 		}
-		//		if confPath != "" && confArg != "" {
-		//			log.Fatal("cannot specify both --config, -c and --confarg, -C")
-		//		}
-		//		if confPath != "" && stdin {
-		//			log.Fatal("cannot specify both --config, -c and --stdin, -n")
-		//		}
-		//		if stdin && confArg != "" {
-		//			log.Fatal("cannot specify both --confarg, -C and --stdin, -n")
-		//		}
-		//		if stdin || confArg != "" || confPath != "" {
 
 		if confPath != "" {
-			//			conf = initConfig()
 			confJSON, err := os.ReadFile(confPath) //nolint
 			if err != nil {
 				log.WithError(err).Fatal("Failed to read config file")
@@ -111,10 +106,58 @@ var surveyCmd = &cobra.Command{
 			survey.ServicesURLs.StunServers = conf.StunServers
 			//survey.DmsgServers = v.dmsgC.ConnectedServersPK()
 		}
+		if dmsgdisc != "" {
+			ipAddr, err = FetchIP(dmsgDisc)
+			if err == nil {
+				survey.IPAddr = ipAddr.String()
+			}
+		}
+
 		s, err := json.MarshalIndent(survey, "", "\t")
 		if err != nil {
 			internal.PrintFatalError(cmd.Flags(), fmt.Errorf("Could not marshal json: %v", err))
 		}
 		fmt.Printf("%s", s)
 	},
+}
+
+// FetchIP fetches the ip address by dmsg servers
+func FetchIP(dmsgDisc string) (string, error) {
+	log := logging.MustGetLogger("ip_skycoin_fetch_dmsg")
+	ctx, cancel := cmdutil.SignalContext(context.Background(), nil)
+	defer cancel()
+
+	pk, sk := cipher.GenerateKeyPair()
+
+	dmsgC, closeDmsg, err := startDmsg(ctx, log, pk, sk, dmsgDisc)
+	if err != nil {
+		return "", fmt.Errorf("failed to start dmsg")
+	}
+	defer closeDmsg()
+
+	ip, err := dmsgC.LookupIP(ctx, nil)
+	return ip.String(), err
+}
+
+func startDmsg(ctx context.Context, log *logging.Logger, pk cipher.PubKey, sk cipher.SecKey, dmsgDisc string) (dmsgC *dmsg.Client, stop func(), err error) {
+	dmsgC = dmsg.NewClient(pk, sk, disc.NewHTTP(dmsgDisc, &http.Client{}, log), &dmsg.Config{MinSessions: dmsg.DefaultMinSessions})
+	go dmsgC.Serve(context.Background())
+
+	stop = func() {
+		err := dmsgC.Close()
+		log.WithError(err).Debug("Disconnected from dmsg network.")
+		fmt.Printf("\n")
+	}
+	log.WithField("public_key", pk.String()).WithField("dmsg_disc", dmsgDisc).
+		Debug("Connecting to dmsg network...")
+
+	select {
+	case <-ctx.Done():
+		stop()
+		return nil, nil, ctx.Err()
+
+	case <-dmsgC.Ready():
+		log.Debug("Dmsg network ready.")
+		return dmsgC, stop, nil
+	}
 }
